# docs/swagger-v2.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/auth/me": {
      "get": {
        "operationId": "AuthController_me",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/AuthUserDto" } }
            }
          },
          "401": { "description": "Нет или истёк access-токен" }
        },
        "security": [{ "bearer": [] }],
        "summary": "Текущий пользователь + профиль",
        "tags": ["auth"]
      }
    },
...
  },
  "info": {
    "title": "Instagram Backend API",
    "description": "NestJS + Prisma + PostgreSQL. Конверт ответа: { data, errors, statusCode }",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": { "scheme": "bearer", "bearerFormat": "JWT", "type": "http", "in": "header" }
    },
    "schemas": {
...
      "AuthUserDto": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "example": "6f1c9e1a-2b3d-4c5e-8a7b-9d0e1f2a3b4c" },
          "userName": { "type": "string", "example": "eraj_dev" },
          "fullName": { "type": "string", "example": "Eraj Karimov" },
          "email": { "type": "string", "example": "eraj@example.com" },
          "phone": { "type": "string", "nullable": true, "example": "+992901234567" },
          "dob": {
            "format": "date-time",
            "type": "string",
            "nullable": true,
            "example": "2000-05-17T00:00:00.000Z"
          },
          "role": { "type": "string", "example": "USER", "enum": ["USER", "ADMIN"] },
          "isPrivate": { "type": "boolean", "example": false },
          "isVerified": { "type": "boolean", "example": false },
          "avatarUrl": { "type": "string", "nullable": true, "description": "Аватар из Profile" }
        },
        "required": ["id", "userName", "fullName", "email", "role", "isPrivate", "isVerified"]
      },
...
    }
  }
}